+++ server/seeds/bannerData.json
@@ -29,7 +29,7 @@
   },
   {
     "name": "Perfect Winter Meal! Warm and Toasty Happy Milk Gacha",
-    "bannerArt": "https://cdn.wikiwiki.jp/to/w/sidem-gstars/%E3%80%90%E3%82%AC%E3%82%B7%E3%83%A3%E3%80%91%E3%80%8C%E4%BB%8A%E5%B9%B4%E3%82%82%E8%89%AF%E3%81%841%E5%B9%B4%E3%81%AB%EF%BC%81%E5%BF%83%E3%82%92%E8%BE%BC%E3%82%81%E3%81%9F%E5%B9%B4%E8%B3%80%E7%8A%B6%E3%82%AC%E3%82%B7%E3%83%A3%E3%80%8D/::ref/banner.png.webp?rev=943e172894cd4de5807c55750ced19f4&t=20230103013705",
+    "bannerArt": "https://cdn.wikiwiki.jp/to/w/sidem-gstars/%E3%80%90%E3%82%AC%E3%82%B7%E3%83%A3%E3%80%91%E3%80%8C%E5%86%AC%E3%81%AE%E9%A3%9F%E5%8D%93%E3%81%AB%E3%82%AA%E3%82%B9%E3%82%B9%E3%83%A1%EF%BC%81%E3%81%BD%E3%81%8B%E3%81%BD%E3%81%8BHappy%20Milk%E3%82%AC%E3%82%B7%E3%83%A3%E3%80%8D/::ref/banner.png.webp?rev=8cbf9e0757bc2242bd66ac9fa2f5cef0&t=20230111152131",
     "backgroundImages": [
       "https://project-imas.wiki/images/7/7f/Kazuki_GS_SSR3%2B.jpg",
       "https://project-imas.wiki/images/d/d2/Kirio_GS_SSR3%2B.jpg"
